fix(*): Fixed defaults for OrderBy #5. dateFrom and dateTo can be set independently of each other.
Updated typedoc, moment, types/node.
Changed default for OrderBy enum to be Relevance.
Some refactoring of query-params. Now always add query-param if it renders to something. In effect more params are included: All enum types  are included by default, plus all primitive types are added when they have value.
Some tests added to verify more on queryconversions.
Docs updated to match changes. 
Fixed so docs is using the master branch instead of the commit-id in the paths, allowing the documentation site to link to the master version of the docs.
Also allowing dateFrom and dateTo to be set independently of each other (previously required boht params to be set).