Added invalid-results dimmer, turned off search-as-you-type
"Search-as-you-type" is challenging to get right. Turning it off for the search UI. Might be better to explore options in the autocomplete dropdown (like Angolia and Google does).
Making the main dimmer whiter (as the other dimmers - and as originally intended).
Added tooltips for reset and search button.
Made the reset-button always show, as it has a purpose after removing the query-text when a search has already been executed.
Changed the invalid check to return false when there has been no previous searches made.
Only managing the invalid-results state from find-service. Should cover more or less all cases. A few edge-cases for categorization, but they are rare. Should be ok.