fix(sample): Set proper caching strategies for the various page resources.
Don't use the cache when loading the index.html 
Make sure that the JS-files at the end of the page are executed in order by removing async.
Don't use the cache when loading the json setttings-files.
Adds "versions" to the urls of local CSS and JS files.
Adds "versions" also to the images referred to in the `ui-settings.json` file.
Adds "version" to the logo in the CSS file.