fix: Allow filters to be added in the settings as a part of the query property on search-client initialization
Allows setting the query.filters property on initialization, thus allowing to set "hidden" filters that will always be included in the calls to the back-end. 
Note that if the filter applied is also a category shown in the category-tree, then it will also be indicated as selected there.
Also made type and enableLogging optional properties in the IAuthenticationSettings