Add cbWarning for "less" severe errors, reserve cbError for fatal errors
On upgrade we recommend adding a `cbWarning` handler for `Find` and `Categorize` as well as making sure that these display to users appropriately (check docs on the `cbWarning` and `cbError` to get a guide on when to use each of them).
In the case of the category's `categoryName` property being null (should have been an array), the search-client will no longer throw an error. Instead the issue is logged in the console as a warning.