feat!: New default `CategorizationType`: `Normal`.
The normal category-filters are not applied server-side for the `CategorizationType.Normal` mode when generating the category-tree. Note that hidden-filters are applied though.
The old default was `CategorizationType.DocumentHitsOnly`, which is now renamed to `CategorizationType.Drilldown`. This mode still works as before, where both normal and hidden category-filters are applied when generating the category-tree server-side.
The `CategorizationType.All` option is untouched and works the same way as before.